Scale back instances in TextShow.FromStringTextShow
This is needed for forward compatibility with
haskell/core-libraries-committee#10, which proposes to add quantified `Show`
superclasses to `Show1` and `Show2`. Unfortunately, several of the instances in
`TextShow.FromStringTextShow` do not mesh well with these new superclasses, so
several of the instances have to either be removed or have their instance
contexts changed as a consequence. It's a bit heavy-handed, but I'm not sure
there exists a better approach. See the `CHANGELOG` for details.

This can be seen as a scaled back version of #55 that only changes instances.
In particular, it does _not_ add quantified `TextShow` superclasses to
`TextShow1` or `TextShow2`. See #56, which discusses that point separately.