Remove outdated Haddock advice in TextShow.Generic
Nowadays, the error message you get when attempting to generically derive a
`TextShow` instance without an accompanying `Generic` instance is much better,
as GHC will simply tell you about the missing `Generic` instance. As such, some
of the advice in the Haddocks in `TextShow.Generic` is now out of date. Let's
just remove it.