Split up GTextShow classes
Currently, all of the `Generic`-related functionality for both `TextShow` and
`TextShow1` is contained within a single set of `GTextShow*` classes. This
proves troublesome when attempting to add a quantified superclass to
`TextShow1`, however, because if we were to define a `TextShow` instance for
`FromGeneric1 f a` (as required by the quantified superclass), then we would
have to add `TextShow a` to the instance context, which is too constrained.

This takes a different approach by splitting `GTextShow*` into two sets of
classes: one `GTextShow*` set (specifically for generic derivation of
`TextShow` functionality) and another `GTextShow1*` set (specifically for
generic derivation of `TextShow1` functionality). This requires some
backwards-incompatible changes to the API in `TextShow.Generic`, but this
change will not affect most users of the code.

This is in preparation for a subsequent commit targeting #56.