TextShow.Generic: Only INLINE things on >= 9.2
The `INLINE` annotations don't cause the compile times to blow up on older
versions of GHC like they do on 8.10, but they do increase noticeably enough
(usually a couple minutes or more on GitHub Actions CI jobs) to the point where
I'm not comfortable with it. Luckily, this doesn't happen on 9.2 or later, so
let's just restrict the use of `INLINE` annotations to those versions.